Organize exceptions
* Create organization hierarchy for exceptions
