+++ common/src/main/java/com/mx/common/accessors/Accessor.java
@@ -9,7 +9,7 @@ import java.util.stream.Collectors;
 
 import javax.annotation.Nonnull;
 
-import com.mx.common.exception.AccessorException;
+import com.mx.common.exception.request.accessor.AccessorException;
 import com.mx.common.gateway.GatewayAPI;
 
 /**
