+++ common/src/main/java/com/mx/common/connect/Request.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 import com.mx.common.collections.MultiValueMap;
 import com.mx.common.collections.MultiValueMappable;
 import com.mx.common.collections.SingleValueMap;
-import com.mx.common.exception.AccessorSystemException;
+import com.mx.common.exception.request.accessor.AccessorSystemException;
 import com.mx.common.request.Feature;
 
 /**
