+++ common/src/main/java/com/mx/common/exception/request/accessor/AccessorException.java
@@ -1,6 +1,7 @@
-package com.mx.common.exception;
+package com.mx.common.exception.request.accessor;
 
 import com.mx.common.accessors.PathResponseStatus;
+import com.mx.common.exception.request.PathRequestException;
 
 /**
  * Base class for exceptions thrown by accessors
