+++ common/src/main/java/com/mx/common/exception/request/accessor/ResourceNotFoundException.java
@@ -1,6 +1,7 @@
-package com.mx.common.exception;
+package com.mx.common.exception.request.accessor;
 
 import com.mx.common.accessors.PathResponseStatus;
+import com.mx.common.exception.request.PathRequestException;
 
 /**
  * Thrown when a resource is requested that does not exist.
