+++ common/src/main/java/com/mx/common/exception/request/accessor/connect/CircuitOpenException.java
@@ -1,4 +1,6 @@
-package com.mx.common.exception;
+package com.mx.common.exception.request.accessor.connect;
+
+import com.mx.common.exception.request.PathRequestException;
 
 /**
  * Thrown on open circuit condition
