+++ common/src/main/java/com/mx/common/exception/request/accessor/connect/UpstreamErrorException.java
@@ -1,9 +1,10 @@
-package com.mx.common.exception;
+package com.mx.common.exception.request.accessor.connect;
 
 import lombok.Getter;
 import lombok.Setter;
 
 import com.mx.common.accessors.PathResponseStatus;
+import com.mx.common.exception.request.accessor.AccessorUserException;
 import com.mx.common.http.HttpStatus;
 
 /**
