+++ common/src/main/java/com/mx/common/exception/request/facility/FacilityException.java
@@ -1,22 +1,23 @@
-package com.mx.common.exception;
+package com.mx.common.exception.request.facility;
 
 import com.mx.common.accessors.PathResponseStatus;
+import com.mx.common.exception.request.PathRequestException;
 
 /**
- * Thrown on unrecoverable error in accessor code
+ * Thrown on unrecoverable error in facility code
  *
  * <p>
  *   See {@link PathRequestException} for usage details
  * </p>
  */
-public class AccessorSystemException extends AccessorException {
-  public AccessorSystemException(String message) {
+public class FacilityException extends PathRequestException {
+  public FacilityException(String message) {
     super(message);
     setReport(true);
     setStatus(PathResponseStatus.INTERNAL_ERROR);
   }
 
-  public AccessorSystemException(String message, Throwable cause) {
+  public FacilityException(String message, Throwable cause) {
     super(message, cause);
     setReport(true);
     setStatus(PathResponseStatus.INTERNAL_ERROR);
