+++ gateway/src/main/java/com/mx/path/gateway/connect/filters/FaultTolerantRequestFilter.java
@@ -3,7 +3,7 @@ package com.mx.path.gateway.connect.filters;
 import com.mx.common.connect.Request;
 import com.mx.common.connect.RequestFilterBase;
 import com.mx.common.connect.Response;
-import com.mx.common.exception.ConnectException;
+import com.mx.common.exception.request.accessor.connect.ConnectException;
 import com.mx.common.http.HttpStatus;
 import com.mx.common.process.FaultTolerantExecutor;
 import com.mx.path.gateway.context.GatewayRequestContext;
