+++ http/src/main/java/com/mx/path/api/connect/http/HttpRequest.java
@@ -2,7 +2,7 @@ package com.mx.path.api.connect.http;
 
 import com.mx.common.connect.Request;
 import com.mx.common.connect.RequestFilter;
-import com.mx.common.exception.AccessorSystemException;
+import com.mx.common.exception.request.accessor.AccessorSystemException;
 
 public class HttpRequest extends Request<HttpRequest, HttpResponse> {
