+++ test-models/src/main/java/com/mx/testing/accessors/IdBaseAccessor.java
@@ -4,7 +4,7 @@ import com.mx.common.accessors.API;
 import com.mx.common.accessors.Accessor;
 import com.mx.common.accessors.AccessorConfiguration;
 import com.mx.common.accessors.AccessorResponse;
-import com.mx.common.exception.AccessorMethodNotImplementedException;
+import com.mx.common.exception.request.accessor.AccessorMethodNotImplementedException;
 import com.mx.common.gateway.GatewayAPI;
 import com.mx.common.gateway.GatewayClass;
 import com.mx.testing.model.Authentication;
